replace `MethodInfoKey` Pair alias with struct to avoid type piracy
Changes `MethodInfoKey` from a `Pair` type alias to a proper struct,
avoiding type piracy that would be introduced by defining
`Pair{...}(::Method)` conversion methods. The struct implements
`Base.iterate` to maintain compatibility with destructuring patterns.

This is a breaking change as it changes the representation of method
info keys, which packages like LCU and Revise depend heavily on. So we
unfotunately need to bump the major version from 2 to 3.

As an alternative, removing `Pair{...}(::Method)` would also be
possible, but that would also break LCU and Revise, so I'm taking this
approach which appears to be a better and moer complete fix.